Move ShareContentButton and share to ui-connected
