+++ packages/apollos-ui-connected/src/utils/index.js
@@ -1,1 +1,2 @@
 export fetchMoreResolver from './fetchMoreResolver';
+export share from './share';
